+++ app/(afterLogin)/createPost/_components/calendar.tsx
@@ -4,8 +4,8 @@ import React, { useState, ChangeEvent, FC } from "react"
 import dayjs, { Dayjs } from "dayjs"
 import advancedFormat from "dayjs/plugin/advancedFormat"
 import isBetween from "dayjs/plugin/isBetween"
-import { generateCalendarOptions, generateDays, renderDayOfWeek, renderDay } from "./calendarUtils"
-import { CalendarProps, DateRange } from "../type"
+import { generateCalendarOptions, generateDays, renderDayOfWeek, renderDay } from "../../../../utils/calendarUtils"
+import { CalendarProps, DateRange } from "./type"
 
 dayjs.extend(advancedFormat)
 dayjs.extend(isBetween)
