+++ components/commons/type.ts
@@ -15,27 +15,4 @@ export type PostCardProps = {
     created_At: Date
 }
 
-//calendar props type
-export type DateRange = {
-    start: Date | null
-    end: Date | null
-}
 
-export type CalendarProps = {
-    onClose: () => void
-}
-
-//form overlay props type
-export type FormOverlayProps = {
-    isContinentOverlayOpen: boolean
-    isCalendarOverlayOpen: boolean
-    onClose: () => void
-    handleContinentSelect: (continent: string) => void
-    selectedContinent: string | null
-}
-
-//form selector props type
-export type FormSelectorProps = {
-    onClick: () => void
-    label: string
-}
